ypalitestinaokruglenie

+++ dlya_testov/sprintf.h
@@ -44,5 +44,5 @@ void handle_e(va_list factor, char *str, int *l, flags *f, int uppercase);
 void handle_o(va_list factor, char *str, int *l, flags *f);
 void handle_p(va_list factor, char *str, int *l, flags *f);
 
-void print_frac_part(char *str, int *l, double num, int frac_len);
-void print_int_part(char *str, int *l, double num);
+void print_frac_part(char *str, int *l, long double num, int frac_len);
+void print_int_part(char *str, int *l, long double num);
